pyproject.toml
[project]
name = "python-fine-tuning-openai"
version = "0.1.0"
description = ""
authors = [
    {name = "Frederico Ferreira",email = "frederico.dida1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
